requirements.txt
# Core dependencies
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.25

# FastAPI and web framework
fastapi==0.108.0
uvicorn[standard]==0.25.0

# Email APIs
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.114.0
msal==1.26.0
requests==2.31.0

# AI/ML
openai==1.7.2

# Scheduling
apscheduler==3.10.4
pytz==2023.3

# Logging and utilities
loguru==0.7.2
python-dotenv==1.0.0

# Database
alembic==1.13.1

# CLI
click==8.1.7
rich==13.7.0

# HTTP client
aiohttp==3.9.1

# Email handling
email-validator==2.1.0
html2text==2020.1.16
beautifulsoup4==4.12.2
lxml==4.9.4

# Date/time handling
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Development (optional)
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0